refactor: move vendor and global jest mocks under test/mock/
