+++ test/mock/@supabase/supabase-js.ts
@@ -1,4 +1,4 @@
-import { createMockSupabaseClient } from '@/lib/jest/mock/@supabase/client';
+import { createMockSupabaseClient } from '@/test/mock/@supabase/client';
 
 jest.mock('@supabase/supabase-js', () => ({
   createClient: jest.fn(() => createMockSupabaseClient()),
