+++ test/mock/globalThis/IntersectionObserver.ts
