+++ test/mock/globalThis/URL/createObjectURL.ts
