+++ test/mock/next/navigation.ts
