+++ tsconfig.json
@@ -49,6 +49,9 @@
       "@/common/*": [
         "./src/common/*"
       ],
+      "@/test/*": [
+        "./test/*"
+      ],
       "@/*": [
         "./src/*"
       ],
